break up codebase, fix bugs, and add individual file option
