+++ app.py
@@ -1,183 +1,13 @@
 import streamlit as st
-import re
-import requests
-from github import Github, GithubException
-from io import StringIO
+from github import Github
+from github_utils import process_github_content
+from ui import render_ui, render_sidebar
 
-
-def parse_github_input(input_line):
-    pr_match = re.match(
-        r"https://github.com/([^/]+/[^/]+)/pull/(\d+)", input_line
-    )
-    if pr_match:
-        return "pr", pr_match.group(1), pr_match.group(2)
-
-    if input_line.startswith("regex:"):
-        return "regex", None, input_line[6:].strip()
-
-    repo_match = re.match(r"https://github.com/([^/]+/[^/]+)/?$", input_line)
-    if repo_match:
-        return "repo", repo_match.group(1), None
-
-    match = re.match(
-        r"https://github.com/([^/]+/[^/]+)(?:/tree/([^/]+))?(/.*)?", input_line
-    )
-    if match:
-        repo_name = match.group(1)
-        branch = match.group(2)
-        path = match.group(3) or ""
-        return "content", repo_name, (path.lstrip("/"), branch)
-
-    return None, None, None
-
-
-def get_default_branch(repo):
-    return repo.default_branch
-
-
-def get_file_content(
-    repo, file_path, regex_patterns, keep_matching, branch=None
-):
-    try:
-        if branch is None:
-            branch = get_default_branch(repo)
-        file_content = repo.get_contents(file_path, ref=branch)
-
-        # Check if the file is too large (over 1MB)
-        if file_content.size > 1000000:
-            return f"File {file_path} is too large to display (size: {file_content.size} bytes)"
-
-        # Handle binary files
-        file_type = file_path.split(".")[-1].lower()
-        if file_type in ["png", "jpg", "jpeg", "gif", "bmp"]:
-            return f"[Binary image file: {file_path}]"
-        elif file_type in ["pdf", "doc", "docx", "xls", "xlsx"]:
-            return f"[Binary document file: {file_path}]"
-
-        # Decode content for text files
-        content = base64.b64decode(file_content.content).decode(
-            "utf-8", errors="replace"
-        )
-
-        lines = content.split("\n")
-        if regex_patterns:
-            if keep_matching:
-                return "\n".join(
-                    line
-                    for line in lines
-                    if any(
-                        re.search(pattern, line) for pattern in regex_patterns
-                    )
-                )
-            else:
-                return "\n".join(
-                    line
-                    for line in lines
-                    if not any(
-                        re.search(pattern, line) for pattern in regex_patterns
-                    )
-                )
-        else:
-            return content
-    except Exception as e:
-        return f"Error: Could not fetch content for {file_path}. {str(e)}"
-
-
-def get_pr_diff(owner, repo, pull_number, token):
-    url = f"https://api.github.com/repos/{owner}/{repo}/pulls/{pull_number}"
-    headers = {
-        "Authorization": f"token {token}",
-        "Accept": "application/vnd.github.v3.diff",
-    }
-    response = requests.get(url, headers=headers)
-    if response.status_code == 200:
-        return response.text
-    else:
-        return f"Error fetching GitHub diff: {response.status_code}"
-
-
-def process_github_content(g, inputs, regex_patterns, keep_matching):
-    output = StringIO()
-    github_token = st.secrets["GITHUB_TOKEN"]
-
-    for input_line in inputs:
-        input_type, repo_name, extra_info = parse_github_input(input_line)
-
-        if input_type is None:
-            output.write(f"\nInvalid input: {input_line}\n")
-            continue
-
-        output.write(f"\n\n--- Content from {input_line} ---\n")
-        try:
-            if input_type == "pr":
-                owner, repo = repo_name.split("/")
-                diff = get_pr_diff(owner, repo, extra_info, github_token)
-                output.write(diff)
-            elif input_type in ["content", "repo"]:
-                repo = g.get_repo(repo_name)
-                default_branch = get_default_branch(repo)
-                if input_type == "repo":
-                    contents = repo.get_contents("", ref=default_branch)
-                    branch = default_branch
-                else:
-                    path, branch = extra_info
-                    branch = branch or default_branch
-                    contents = repo.get_contents(path, ref=branch)
-                if not isinstance(contents, list):
-                    contents = [contents]
-                while contents:
-                    file_content = contents.pop(0)
-                    if file_content.type == "dir":
-                        contents.extend(
-                            repo.get_contents(file_content.path, ref=branch)
-                        )
-                    else:
-                        output.write(f"\n--- {file_content.path} ---\n\n")
-                        content = get_file_content(
-                            repo,
-                            file_content.path,
-                            regex_patterns,
-                            keep_matching,
-                            branch,
-                        )
-                        output.write(content)
-                        output.write("\n")
-            elif input_type == "regex":
-                for repo in g.get_user().get_repos():
-                    default_branch = get_default_branch(repo)
-                    for content in repo.get_contents("", ref=default_branch):
-                        if re.match(extra_info, content.path):
-                            output.write(
-                                f"\n--- {repo.full_name}/{content.path} ---\n\n"
-                            )
-                            content = get_file_content(
-                                repo,
-                                content.path,
-                                regex_patterns,
-                                keep_matching,
-                                default_branch,
-                            )
-                            output.write(content)
-                            output.write("\n")
-        except GithubException as e:
-            if e.status == 404:
-                output.write(
-                    f"Error: Repository, path, or PR not found. Please check the URL and ensure you have access to this repository.\n"
-                )
-            else:
-                output.write(
-                    f"GitHub API error: {e.status} - {e.data.get('message', 'Unknown error')}\n"
-                )
-        except Exception as e:
-            output.write(f"An error occurred: {str(e)}\n")
-    return output.getvalue()
+# Set page config at the very beginning
+st.set_page_config(page_title="GitHub Stitcher", page_icon="🧵", layout="wide")
 
 
 def main():
-    st.set_page_config(
-        page_title="GitHub Stitcher", page_icon="🧵", layout="wide"
-    )
-
     st.title("🧵 GitHub Stitcher")
     st.markdown(
         """
@@ -187,31 +17,9 @@ def main():
     """
     )
 
-    # Use GitHub token from Streamlit secrets
     github_token = st.secrets["GITHUB_TOKEN"]
 
-    # Input for GitHub URLs and regex patterns
-    github_inputs = st.text_area(
-        "Enter GitHub URLs, PR links, or regex patterns (one per line):",
-        height=150,
-        help="Enter URLs to files, folders, PRs, or use 'regex:' prefix for file path patterns. "
-        "e.g., https://github.com/username/repo/tree/branch/folder, "
-        "https://github.com/username/repo/pull/123, or regex:.*\\.py$",
-    )
-
-    # Input for text filtering
-    regex_patterns = st.text_area(
-        "Enter regex patterns for text filtering (one per line):",
-        height=100,
-        help="These patterns will be used to filter the content of the files. Leave empty to include all content.",
-    )
-
-    # Radio button to choose between keeping or omitting matched lines
-    filter_mode = st.radio(
-        "Choose filtering mode:",
-        ("Keep matching lines", "Omit matching lines"),
-        help="Select whether to keep or omit the lines that match the regex patterns above.",
-    )
+    github_inputs, regex_patterns, filter_mode = render_ui()
 
     if st.button(
         "🧵 Stitch Content", help="Click to fetch and stitch the content"
@@ -234,12 +42,12 @@ def main():
 
         with st.spinner("Fetching and stitching content..."):
             g = Github(github_token)
-            all_content = process_github_content(
+            all_content, error_occurred = process_github_content(
                 g, inputs, patterns, keep_matching
             )
 
-            if "Error:" in all_content:
-                st.error(
+            if error_occurred:
+                st.warning(
                     "⚠️ Some errors occurred while fetching content. Please check the output below."
                 )
             else:
@@ -255,19 +63,7 @@ def main():
                 help="Click to download the stitched content as a text file",
             )
 
-    st.sidebar.title("About")
-    st.sidebar.info(
-        """
-        GitHub Stitcher was created by [Max Ghenis](mailto:mghenis@gmail.com).
-
-        This tool allows you to:
-        - Fetch content from GitHub files, folders, or entire repositories
-        - Fetch diffs from Pull Requests
-        - Use regex patterns to filter specific file paths
-        - Keep or omit lines based on regex patterns
-        - Combine the filtered content into a single, easy-to-use output
-        """
-    )
+    render_sidebar()
 
 
 if __name__ == "__main__":
